pyproject.toml
[project]
name = "eho-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "chekh",email = "chechirkinbogd@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.19.0,<4.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
